Refactor BGP Speaker tests into a different module since they're long
