+++ crates/bgp-speaker/src/lib.rs
@@ -29,109 +29,4 @@ pub mod peer;
 pub mod supervisor;
 
 #[cfg(test)]
-mod peer_test;
-
-#[cfg(test)]
-pub mod test {
-    use async_trait::async_trait;
-    use std::{io, io::Cursor, net::SocketAddr, time::Duration};
-
-    use tokio_test::io::{Builder, Mock};
-
-    use crate::{codec::BgpCodec, connection::ActiveConnect};
-    use netgauze_bgp_pkt::BgpMessage;
-    use netgauze_parse_utils::WritablePdu;
-
-    /// Wrap [Builder] allowing it to accept BgpMessages for read and write
-    /// mocks rather than `&[u8]`.
-    #[derive(Default, Debug)]
-    pub struct BgpIoMockBuilder {
-        io_builder: Builder,
-    }
-
-    impl BgpIoMockBuilder {
-        pub fn new() -> Self {
-            Self {
-                io_builder: Builder::new(),
-            }
-        }
-
-        /// See [Builder::read]
-        pub fn read(&mut self, msg: BgpMessage) -> &mut Self {
-            let buf = vec![];
-            let mut cursor = Cursor::new(buf);
-            msg.write(&mut cursor).unwrap();
-            self.io_builder.read(&cursor.into_inner());
-            self
-        }
-
-        /// See [Builder::read]
-        pub fn read_u8(&mut self, buf: &[u8]) -> &mut Self {
-            self.io_builder.read(buf);
-            self
-        }
-
-        /// See [Builder::write]
-        pub fn write(&mut self, msg: BgpMessage) -> &mut Self {
-            let buf = vec![];
-            let mut cursor = Cursor::new(buf);
-            msg.write(&mut cursor).unwrap();
-            self.io_builder.write(&cursor.into_inner());
-            self
-        }
-
-        /// See [Builder::write]
-        pub fn write_u8(&mut self, buf: &[u8]) -> &mut Self {
-            self.io_builder.write(buf);
-            self
-        }
-
-        /// See [Builder::wait]
-        pub fn wait(&mut self, duration: Duration) -> &mut Self {
-            self.io_builder.wait(duration);
-            self
-        }
-
-        /// See [Builder::build]
-        pub fn build(&mut self) -> Mock {
-            self.io_builder.build()
-        }
-    }
-
-    pub struct MockActiveConnect {
-        pub peer_addr: SocketAddr,
-        pub io_builder: BgpIoMockBuilder,
-        pub connect_delay: Duration,
-    }
-
-    #[async_trait]
-    impl ActiveConnect<SocketAddr, Mock, BgpCodec> for MockActiveConnect {
-        async fn connect(&mut self, peer_addr: SocketAddr) -> io::Result<Mock> {
-            assert_eq!(self.peer_addr, peer_addr);
-            if !self.connect_delay.is_zero() {
-                tokio::time::sleep(self.connect_delay).await;
-            }
-            Ok(self.io_builder.build())
-        }
-    }
-
-    /// An [ActiveConnect] that always fail to make a connection
-    pub struct MockFailedActiveConnect {
-        pub peer_addr: SocketAddr,
-        pub connect_delay: Duration,
-    }
-
-    #[async_trait]
-    impl ActiveConnect<SocketAddr, Mock, BgpCodec> for MockFailedActiveConnect {
-        async fn connect(&mut self, peer_addr: SocketAddr) -> io::Result<Mock> {
-            assert_eq!(self.peer_addr, peer_addr);
-            if !self.connect_delay.is_zero() {
-                tokio::time::sleep(self.connect_delay).await;
-            }
-            Err(io::Error::new(
-                io::ErrorKind::ConnectionRefused,
-                "MockFailedActiveConnect connection refused",
-            ))
-        }
-    }
-}
+mod tests;
