+++ crates/bgp-speaker/src/supervisor.rs
@@ -165,87 +165,3 @@ impl<
         Ok((rx, peer_handle))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::connection::TcpActiveConnect;
-    use std::net::{SocketAddr, SocketAddrV4};
-
-    #[tokio::test]
-    async fn test_add_remove_peers() -> Result<(), PeersSupervisorError> {
-        let my_bgp_id = Ipv4Addr::new(1, 1, 1, 1);
-        let peer_bgp_id = Ipv4Addr::new(1, 1, 1, 1);
-        let my_asn = 100;
-        let peer_asn = 200;
-        let peer_addr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(192, 168, 1, 1), 179));
-        let peer_properties = PeerProperties::new(
-            my_asn,
-            peer_asn,
-            my_bgp_id,
-            peer_bgp_id,
-            peer_addr,
-            false,
-            false,
-        );
-        let mut supervisor = PeersSupervisor::new(my_asn, my_bgp_id);
-
-        let (_rx, _peer_handle) = supervisor.create_peer(
-            peer_addr.ip(),
-            peer_properties,
-            PeerConfig::default(),
-            TcpActiveConnect,
-            EchoCapabilitiesPolicy::new(my_asn, false, my_bgp_id, 100, Vec::new(), Vec::new()),
-        )?;
-        let second_create = supervisor.create_peer(
-            peer_addr.ip(),
-            peer_properties,
-            PeerConfig::default(),
-            TcpActiveConnect,
-            EchoCapabilitiesPolicy::new(my_asn, false, my_bgp_id, 100, Vec::new(), Vec::new()),
-        );
-        let removed_peer = supervisor.remove_peer(&peer_addr.ip());
-        let non_existing_peer = supervisor.remove_peer(&peer_addr.ip());
-
-        assert_eq!(second_create.err(), Some(PeersSupervisorError::PeerExists));
-        assert!(removed_peer.is_some());
-        assert!(non_existing_peer.is_none());
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_dynamic_peers() -> Result<(), PeersSupervisorError> {
-        let my_bgp_id = Ipv4Addr::new(1, 1, 1, 1);
-        let peer_bgp_id = Ipv4Addr::new(1, 1, 1, 1);
-        let my_asn = 100;
-        let peer_asn = 200;
-        let peer_addr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::new(192, 168, 1, 1), 179));
-        let peer_properties = PeerProperties::new(
-            my_asn,
-            peer_asn,
-            my_bgp_id,
-            peer_bgp_id,
-            peer_addr,
-            false,
-            false,
-        );
-        let mut supervisor = PeersSupervisor::new(my_asn, my_bgp_id);
-
-        let (_rx, _peer_handle) =
-            supervisor.dynamic_peer(peer_addr.ip(), peer_addr, TcpActiveConnect)?;
-        let second_create = supervisor.create_peer(
-            peer_addr.ip(),
-            peer_properties,
-            PeerConfig::default(),
-            TcpActiveConnect,
-            EchoCapabilitiesPolicy::new(my_asn, false, my_bgp_id, 100, Vec::new(), Vec::new()),
-        );
-
-        let removed_peer = supervisor.remove_peer(&peer_addr.ip());
-        let non_existing_peer = supervisor.remove_peer(&peer_addr.ip());
-        assert_eq!(second_create.err(), Some(PeersSupervisorError::PeerExists));
-        assert!(removed_peer.is_some());
-        assert!(non_existing_peer.is_none());
-        Ok(())
-    }
-}
